requirements.txt
google-cloud-storage
google-cloud-bigquery
scikit-learn
joblib
pandas